Configures capturer service with Supervisord
Migrates the packet capture process to be managed by Supervisord.

This change provides better process management for the capturer service,
ensuring it restarts automatically if it crashes and simplifies the overall
service configuration. It also adds a healthcheck that verifies that tshark is
dumping pcap files.

The Dockerfile is updated to install Supervisord, Tshark and its dependencies and
configure it with a configuration file that manages the tshark process.